backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary
openpyxl==3.1.5
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
slowapi==0.1.9
google-genai==1.8.0
scalar-fastapi==1.8.0
